site/requirements.txt
markdown
jinja2
python-frontmatter
requests
typer
